Fixed constraints & bioLabel

+++ p2p/LoginViewController.swift
@@ -42,7 +42,7 @@ class LoginViewController: UIViewController {
 extension LoginViewController: UITextFieldDelegate {
     func textFieldDidBeginEditing(_ textField: UITextField) {
         self.view.layoutIfNeeded()
-        logoTopConstraint.constant = 0 - 49 - 20
+        logoTopConstraint.constant = CGFloat(0 - 49 - 20)
         UIView.animate(withDuration: 0.3) {
             self.view.layoutIfNeeded()
         }
